Remove the old *_data properties
To support both file names and inline data (with temp files, which are deleted
at the end) would be much more complicated, and we don't need it for now.

NOTE: this is a bit messy, some of the old data stuff is still there (but
probably doesn't work).  Not production-quality code!